[Android] Expose embedding apis for video fullscreen.
The onShow/HideCustomView() was implemented in Crosswalk by default and
not exposed, it could not be overrided by customer to implement the video
fullscreen.
This patch is to expose onShow/HideCustomView().

BUG=XWALK-6939
