+++ runtime/android/core_internal/src/org/xwalk/core/internal/CustomViewCallbackInternal.java
@@ -1,0 +1,20 @@
+// Copyright (c) 2016 Intel Corporation. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+package org.xwalk.core.internal;
+
+/**
+ * A callback interface used by the host application to notify
+ * the current page that its custom view has been dismissed.
+ */
+@XWalkAPI(instance = CustomViewCallbackHandlerInternal.class)
+public interface CustomViewCallbackInternal {
+    /**
+     * Invoked when the host application dismisses the
+     * custom view.
+     * @since 7.0
+     */
+    @XWalkAPI
+    public void onCustomViewHidden();
+}
